third_party/WebKit/LayoutTests/http/tests/devtools/elements/attribute-modified-ns-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that elements panel updates dom tree structure upon changing the attribute with namespace.


Running: testDumpInitial
========= Original ========
  <a id="node" xlink:href="http://localhost">link</a>

Running: testAttributeUpdated
===== On attribute set =====
  <a id="node" xlink:href="changed-url">link</a>

Running: testAttributeRemoved
=== On attribute removed ===
  <a id="node">link</a>
